Compatibility with 64-bit OS 	modified:   VDIF_sim.c 	modified:   cuda_fft_xspec.cu 	new file:   hidoi 	new file:   hidoi.c 	modified:   shm_access.c 	modified:   shm_init_create.c

+++ shm_access.c
@@ -14,7 +14,7 @@ int shm_access(
 	key_t	shm_key,		// INPUT:   Keyword for Shared Memory
 	size_t	shm_size,		// INPUT:   Size of the SHM [bytes]
 	int		*shrd_id,		// OUTPUT:  Pointer to the Shared Memory ID
-	int		*shm_ptr)		// OUTPUT:  Pointer to the Top of SHM
+	long	*shm_ptr)		// OUTPUT:  Pointer to the Top of SHM
 {
 /*
 -------------------------------------------------- OPEN SHARED MEMORY
@@ -25,7 +25,7 @@ int shm_access(
 		printf("Can't Access to the Shared Memory !! \n" );
 		return(-1);
 	}
-	*shm_ptr = (int)shmat( *shrd_id, NULL, 0);
+	*shm_ptr = (long)shmat( *shrd_id, NULL, 0);
 
 	//-------- ENDING --------
 	return( *shrd_id);
